[CI][AMD] Isolate pipeline generator tests
